Admin

+++ backend/controllers/users.js
@@ -291,36 +291,34 @@ const getUserById = async (req, res) => {
 };
 
 const addToBasket = async (req, res) => {
-  try {
-    const { user_id } = req.token;
-    const { product_id, quantity } = req.body;
-    const result = await client.query(
-      `SELECT *, order_items.id FROM order_items WHERE product_id=$1 AND user_id=$2`,
-      [product_id, user_id]
-    );
+    try {
+        const { user_id } = req.token;
+        const { product_id, quantity } = req.body;
+        const result = await client.query(`SELECT *, order_items.id FROM order_items WHERE product_id=$1 AND user_id=$2 AND is_deleted=0`, [product_id, user_id]);
+        if (result.rows.length) {
+            await client.query(`UPDATE order_items SET quantity = $1 WHERE product_id = $2 AND user_id=$3 AND is_deleted=0 RETURNING *`, [quantity, product_id, user_id]);
+            const result = await client.query(`SELECT *,order_items.id FROM order_items INNER JOIN products ON products.id = $1 AND order_items.product_id=$1 WHERE user_id = $2 AND order_items.is_deleted=0`, [product_id, user_id]);
+            const remainingItems = await client.query(`SELECT *, order_items.id FROM order_items INNER JOIN products ON order_items.product_id = products.id WHERE user_id=$1 AND order_items.is_deleted=0`, [user_id]);
+            res.status(200).json({
+                success: true,
+                result: result.rows,
+                remaining: remainingItems.rows
+            });
+        } else {
+            const update = await client.query(`INSERT INTO order_items (quantity, product_id, user_id) VALUES ($1,$2,$3) RETURNING *`, [quantity, product_id, user_id]);
+            res.status(200).json({
+                success: true,
+                result: update.rows
+            });
+        }
+    } catch (error) {
+        console.log(error.message);
+        res.json({
+            success: false,
+            message: "Server Error",
+            error: error.message
+        });
 
-    if (result.rows.length) {
-      await client.query(
-        `UPDATE order_items SET quantity = $1 WHERE product_id = $2 AND user_id=$3 RETURNING *`,
-        [quantity, product_id, user_id]
-      );
-      const result = await client.query(
-        `SELECT *,order_items.id FROM order_items INNER JOIN products ON products.id = $1 AND order_items.product_id=$1 WHERE user_id = $2`,
-        [product_id, user_id]
-      );
-      res.status(200).json({
-        success: true,
-        result: result.rows,
-      });
-    } else {
-      const update = await client.query(
-        `INSERT INTO order_items (quantity, product_id, user_id) VALUES ($1,$2,$3) RETURNING *`,
-        [quantity, product_id, user_id]
-      );
-      res.status(200).json({
-        success: true,
-        result: update.rows,
-      });
     }
   } catch (error) {
     console.log(error.message);
@@ -333,44 +331,40 @@ const addToBasket = async (req, res) => {
 };
 
 const getUserBasket = async (req, res) => {
-  try {
-    const { user_id } = req.token;
-    const result = await client.query(
-      `SELECT *, order_items.id FROM order_items INNER JOIN products ON order_items.product_id = products.id WHERE user_id=$1`,
-      [user_id]
-    );
-    res.json({
-      success: true,
-      result: result.rows,
-    });
-  } catch (error) {
-    res.json({
-      success: false,
-      error: error.message,
-    });
-  }
-};
+    try {
+        const { user_id } = req.token;
+        const result = await client.query(`SELECT *, order_items.id FROM order_items INNER JOIN products ON order_items.product_id = products.id WHERE user_id=$1 AND order_items.is_deleted=0`, [user_id]);
+        res.json({
+            success: true,
+            result: result.rows
+        });
+    } catch (error) {
+        res.json({
+            success: false,
+            error: error.message
+        });
+    }
+}
 
 const deleteCartItem = async (req, res) => {
-  try {
-    const { itemId } = req.params;
-    const { user_id } = req.token;
-    const result = await client.query(
-      `DELETE FROM order_items WHERE user_id=$1 AND id=$2 RETURNING *`,
-      [user_id, itemId]
-    );
-    res.json({
-      success: true,
-      result: result.rows[0],
-    });
-  } catch (error) {
-    console.log(error.message);
-    res.json({
-      success: true,
-      error: error.message,
-    });
-  }
-};
+    try {
+        const { itemId } = req.params;
+        const { user_id } = req.token;
+        const result = await client.query(`DELETE FROM order_items WHERE user_id=$1 AND id=$2 RETURNING *`, [user_id, itemId]);
+        const remainingItems = await client.query(`SELECT *, order_items.id FROM order_items INNER JOIN products ON order_items.product_id = products.id WHERE user_id=$1 AND order_items.is_deleted=0`, [user_id]);
+        res.json({
+            success: true,
+            result: result.rows[0],
+            remaining: remainingItems.rows
+        })
+    } catch (error) {
+        console.log(error.message);
+        res.json({
+            success: true,
+            error: error.message
+        })
+    }
+}
 
 module.exports = {
   userRegister,
